+++ doc/CHANGES.rst
@@ -1,2 +1,17 @@
 CHANGES LOG
 ===========
+
+0.1
+---
+
+Main version of AnyBlok. You can with this version
+
+* Create your own application
+* Connect on a database
+* Define blok
+* Define Column types
+* Define Relationship types
+* Define Model (SQL or not)
+* Define SQL view
+* Define more than one Model on a spécific table
+* Write unittest for your blok